Added Damage Visual, Monster Data
There is Monster data that has stats for how damage is done. Also have visual for damage number. Need to redo some UI as it is not scaling with phone correct. Need to have attacks have a base attack % that they may miss and set dodge base to 0 or just remove.